refactor(kornia-apriltag): consolidate SIMD kernels into an ops/ module
Replace the thin simd.rs (feature-probe only) with an ops/ module that houses the
dispatched kernels per architecture:

- ops/scalar.rs, ops/neon.rs, ops/avx2.rs — the classify_row, fill_tile_stats, and
  smooth_interior kernels (+ shared tile_min_max), one file per ISA.
- ops/mod.rs — runtime feature detection (has_avx2 / has_avx2_fma, formerly simd.rs)
  plus the NEON/AVX2/scalar dispatchers; call sites now use crate::ops::*.

The kernels that are genuinely interleaved with scalar algorithm logic
(gradient-cluster scan, line-fit error, RLE run extension) stay in place but select
their path through crate::ops::has_avx2(). Behavior is unchanged: C-vs-kornia
parity, the AVX2-vs-scalar unit parity tests, and pose tests all pass on x86_64 and
aarch64; the 9 pre-existing aarch64-golden lib failures are unaffected.
