+++ crates/kornia-apriltag/src/lib.rs
@@ -21,8 +21,8 @@ use crate::{
 /// Error types for AprilTag detection.
 pub mod errors;
 
-/// Runtime SIMD feature detection and dispatch helpers.
-pub mod simd;
+/// Architecture-dispatched SIMD kernels (NEON / AVX2 / scalar) + feature detection.
+pub mod ops;
 
 /// Utility functions for AprilTag detection.
 pub mod utils;
